+++ unit/Luni/Component/MagentoDriver/SchemaBuilder/Table/AttributeOptionValue.php
@@ -30,8 +30,8 @@ class AttributeOptionValue
     {
         $table = $this->schema->createTable('eav_attribute_option_value');
         
-        $table->addColumn('value_id', 'smallint', ['unsigned' => true, 'autoincrement' => true]);
-        $table->addColumn('option_id', 'smallint', ['unsigned' => true, 'default' => 0]);
+        $table->addColumn('value_id', 'integer', ['unsigned' => true, 'autoincrement' => true]);
+        $table->addColumn('option_id', 'integer', ['unsigned' => true, 'default' => 0]);
         $table->addColumn('store_id', 'smallint', ['unsigned' => true, 'default' => 0]);
         $table->addColumn('value', 'string', ['length' => 255, 'notnull' => false]);
